requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.1.27
langchain-ollama>=0.0.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.5.1
pydantic>=2.0.0
langchain-text-splitters>=0.0.1
unstructured>=0.10.30
python-magic>=0.4.27
numpy>=1.26.2 